⚡ Bolt: Optimize plant filtering logic
- Extract filtering logic to src/lib/plantFilter.ts
- Optimize Set intersection by iterating smaller set
- Optimize subset checks with size guard
- Move PreparedPlant type to src/types/plant.ts
